+++ platformclientv2/analyticssession.go
@@ -40,6 +40,10 @@ type Analyticssession struct {
 	AgentBullseyeRing *int `json:"agentBullseyeRing,omitempty"`
 
 
+	// AgentOwned - Flag indicating an agent-owned callback
+	AgentOwned *bool `json:"agentOwned,omitempty"`
+
+
 	// Ani - Automatic Number Identification (caller's number)
 	Ani *string `json:"ani,omitempty"`
 
@@ -236,6 +240,10 @@ type Analyticssession struct {
 	SkipEnabled *bool `json:"skipEnabled,omitempty"`
 
 
+	// TimeoutSeconds - The number of seconds before PureCloud begins the call for a call back (0 disables automatic calling)
+	TimeoutSeconds *int `json:"timeoutSeconds,omitempty"`
+
+
 	// UsedRouting - Complete routing method
 	UsedRouting *string `json:"usedRouting,omitempty"`
 
@@ -266,10 +274,6 @@ type Analyticssession struct {
 
 	// Segments - List of segments for this session
 	Segments *[]Analyticsconversationsegment `json:"segments,omitempty"`
-
-
-	// TimeoutSeconds - The number of seconds before PureCloud begins the call for a call back (0 disables automatic calling)
-	TimeoutSeconds *int `json:"timeoutSeconds,omitempty"`
 
 }